_get_data_frame is no longer used by runtime code. Move it to legacy_add_rows_test.py, and remove tests against it
