+++ lib/tests/streamlit/legacy_add_rows_test.py
@@ -13,14 +13,16 @@
 # limitations under the License.
 
 """Unit test of dg._legacy_add_rows()."""
+from typing import Optional
 
 import pandas as pd
 import pyarrow as pa
 
 from streamlit.errors import StreamlitAPIException
+from streamlit.proto.DataFrame_pb2 import DataFrame
+from streamlit.proto.Delta_pb2 import Delta
 from streamlit.script_run_context import get_script_run_ctx
 import streamlit as st
-import streamlit.elements.legacy_data_frame as data_frame
 from tests import testutil
 
 
@@ -78,7 +80,7 @@ class DeltaGeneratorAddRowsTest(testutil.DeltaGeneratorTestCase):
             el = delta(DATAFRAME)
             el._legacy_add_rows(NEW_ROWS)
 
-            df_proto = data_frame._get_data_frame(self.get_delta_from_queue())
+            df_proto = _get_data_frame(self.get_delta_from_queue())
 
             # Test that the add_rows delta is properly melted
             rows = df_proto.data.cols[0].int64s.data
@@ -93,7 +95,7 @@ class DeltaGeneratorAddRowsTest(testutil.DeltaGeneratorTestCase):
             el = method(DATAFRAME)
 
             # Make sure it has 2 rows in it.
-            df_proto = data_frame._get_data_frame(self.get_delta_from_queue())
+            df_proto = _get_data_frame(self.get_delta_from_queue())
             num_rows = len(df_proto.data.cols[0].int64s.data)
             self.assertEqual(2, num_rows)
 
@@ -101,7 +103,7 @@ class DeltaGeneratorAddRowsTest(testutil.DeltaGeneratorTestCase):
             el._legacy_add_rows(NEW_ROWS)
 
             # Make sure the add_rows proto looks like we expect.
-            df_proto = data_frame._get_data_frame(self.get_delta_from_queue())
+            df_proto = _get_data_frame(self.get_delta_from_queue())
             rows = df_proto.data.cols[0].int64s.data
             self.assertEqual([3, 4, 5], rows)
 
@@ -118,7 +120,7 @@ class DeltaGeneratorAddRowsTest(testutil.DeltaGeneratorTestCase):
             el = method(DATAFRAME_WITH_INDEX)
 
             # Make sure it has 2 rows in it.
-            df_proto = data_frame._get_data_frame(self.get_delta_from_queue())
+            df_proto = _get_data_frame(self.get_delta_from_queue())
             num_rows = len(df_proto.data.cols[0].int64s.data)
             self.assertEqual(2, num_rows)
 
@@ -126,7 +128,7 @@ class DeltaGeneratorAddRowsTest(testutil.DeltaGeneratorTestCase):
             el._legacy_add_rows(NEW_ROWS_WITH_INDEX)
 
             # Make sure the add_rows proto looks like we expect.
-            df_proto = data_frame._get_data_frame(self.get_delta_from_queue())
+            df_proto = _get_data_frame(self.get_delta_from_queue())
             rows = df_proto.data.cols[0].int64s.data
             self.assertEqual([30, 40, 50], rows)
 
@@ -144,13 +146,13 @@ class DeltaGeneratorAddRowsTest(testutil.DeltaGeneratorTestCase):
         for method in all_methods:
             # Create a new data-carrying element (e.g. st._legacy_dataframe)
             el = method(None)
-            data_frame._get_data_frame(self.get_delta_from_queue())
+            _get_data_frame(self.get_delta_from_queue())
 
             # This is what we're testing:
             el._legacy_add_rows(DATAFRAME_WITH_INDEX)
 
             # Make sure there are 2 rows in it now.
-            df_proto = data_frame._get_data_frame(self.get_delta_from_queue())
+            df_proto = _get_data_frame(self.get_delta_from_queue())
             num_rows = len(df_proto.data.cols[0].int64s.data)
             self.assertEqual(2, num_rows)
 
@@ -165,13 +167,13 @@ class DeltaGeneratorAddRowsTest(testutil.DeltaGeneratorTestCase):
         for method in all_methods:
             # Create a new data-carrying element (e.g. st._legacy_dataframe)
             el = method(None)
-            data_frame._get_data_frame(self.get_delta_from_queue())
+            _get_data_frame(self.get_delta_from_queue())
 
             # This is what we're testing:
             el._legacy_add_rows(DATAFRAME)
 
             # Make sure there are 2 rows in it now.
-            df_proto = data_frame._get_data_frame(self.get_delta_from_queue())
+            df_proto = _get_data_frame(self.get_delta_from_queue())
             num_rows = len(df_proto.data.cols[0].int64s.data)
             self.assertEqual(2, num_rows)
 
@@ -188,7 +190,7 @@ class DeltaGeneratorAddRowsTest(testutil.DeltaGeneratorTestCase):
             el = method(DATAFRAME)
 
             # Make sure it has 2 rows in it.
-            df_proto = data_frame._get_data_frame(self.get_delta_from_queue())
+            df_proto = _get_data_frame(self.get_delta_from_queue())
             num_rows = len(df_proto.data.cols[0].int64s.data)
             self.assertEqual(2, num_rows)
 
@@ -212,7 +214,7 @@ class DeltaGeneratorAddRowsTest(testutil.DeltaGeneratorTestCase):
             el = method(DATAFRAME)
 
             # Make sure it has 2 rows in it.
-            df_proto = data_frame._get_data_frame(self.get_delta_from_queue())
+            df_proto = _get_data_frame(self.get_delta_from_queue())
             num_rows = len(df_proto.data.cols[0].int64s.data)
             self.assertEqual(2, num_rows)
 
@@ -220,7 +222,7 @@ class DeltaGeneratorAddRowsTest(testutil.DeltaGeneratorTestCase):
             el._legacy_add_rows(mydata1=NEW_ROWS)
 
             # Make sure the add_rows proto looks like we expect
-            df_proto = data_frame._get_data_frame(self.get_delta_from_queue(), name="mydata1")
+            df_proto = _get_data_frame(self.get_delta_from_queue(), name="mydata1")
             rows = df_proto.data.cols[0].int64s.data
             self.assertEqual([3, 4, 5], rows)
 
@@ -235,7 +237,7 @@ class DeltaGeneratorAddRowsTest(testutil.DeltaGeneratorTestCase):
             el = method(DATAFRAME)
 
             # Make sure it has 2 rows in it.
-            df_proto = data_frame._get_data_frame(self.get_delta_from_queue())
+            df_proto = _get_data_frame(self.get_delta_from_queue())
             num_rows = len(df_proto.data.cols[0].int64s.data)
             self.assertEqual(2, num_rows)
 
@@ -304,3 +306,53 @@ class DeltaGeneratorAddRowsTest(testutil.DeltaGeneratorTestCase):
             # Clear the queue so the next loop is like a brand new test.
             get_script_run_ctx().reset()
             self.forward_msg_queue.clear()
+
+
+def _get_data_frame(delta: Delta, name: Optional[str] = None) -> DataFrame:
+    """Extract the dataframe protobuf from a delta protobuf."""
+    delta_type = delta.WhichOneof("type")
+
+    if delta_type == "new_element":
+        element_type = delta.new_element.WhichOneof("type")
+
+        # Some element types don't support named datasets.
+        if name and element_type in ("data_frame", "table", "chart"):
+            raise ValueError("Dataset names not supported for st.%s" % element_type)
+
+        if element_type in "data_frame":
+            return delta.new_element.data_frame
+        elif element_type in "table":
+            return delta.new_element.table
+        elif element_type == "chart":
+            return delta.new_element.chart.data
+        elif element_type == "vega_lite_chart":
+            chart_proto = delta.new_element.vega_lite_chart
+            if name:
+                return _get_or_create_dataset(chart_proto.datasets, name)
+            elif len(chart_proto.datasets) == 1:
+                # Support the case where the dataset name was randomly given by
+                # the charting library (e.g. Altair) and the user has no
+                # knowledge of it.
+                return chart_proto.datasets[0].data
+            else:
+                return chart_proto.data
+        # TODO: Support DeckGL. Need to figure out how to handle layer indices
+        # first.
+
+    elif delta_type == "add_rows":
+        if delta.add_rows.has_name and name != delta.add_rows.name:
+            raise ValueError('No dataset found with name "%s".' % name)
+        return delta.add_rows.data
+    else:
+        raise ValueError("Cannot extract DataFrame from %s." % delta_type)
+
+
+def _get_or_create_dataset(datasets_proto, name):
+    for dataset in datasets_proto:
+        if dataset.has_name and dataset.name == name:
+            return dataset.data
+
+    dataset = datasets_proto.add()
+    dataset.name = name
+    dataset.has_name = True
+    return dataset.data
